Improve type handling for PrimaryKeyRelatedField
The existing code for handling a PrimaryKeyRelatedField in a serializer is
to hardcode the primary key to the union type `number | string`.

This has two disadvantages:

1) It is possible to look up which field type is being used as the
primary key on the related model and so it is possible determine more
narrowly which typescript type is appropriate. For example, if the primary
key field on the related model is a number, then ideally we want the
resulting typescript type to be `number` rather than `number | string`.
Likewise, if it were a string, then ideally we'd want the resulting
typescript type to be `string` rather than `number | string`.

2) In some cases, the most appropriate typescript for a Django primary
key might be neither number nor string. In principle, it could be a
`File``. (Indeed, in principle it could be a `boolean`, though I imagine
there aren't many real life cases of that.)

In this patch, instead of hardcoding the typescript type of a
PrimaryKeyRelatedField, we determine what kind of field is being used as
the primary key on the related model and then use the existing logic to
determine the most appropriate typescript type.